# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Trailstamp Notes API
  version: "4.2.0"
  license:
    name: Proprietary
    url: https://github.com/kevin861005/HapeeTrail
  description: |
    第一階段便條 API。業務端點全部由 **HapeeTrail 服務**（Spring Boot，東京）提供，
    路徑為資源導向的 REST。

    **兩個 base URL，別混**：
    - 業務 API → 本檔頂層的 `servers`（HapeeTrail 服務）。
    - `/auth/v1/*`（匿名登入、刷新、登出、日後的帳號綁定）→ **Supabase GoTrue**，
      見 `/auth/v1/signup` 這個 path 自己的 `servers`。服務不代理任何 auth 路徑。

    **登出走 Supabase（`POST /auth/v1/logout`，各語言 SDK 的 signOut），本服務沒有登出端點。**
    被登出的 session 已簽發、尚未過期的 access token 打本服務一律 401（見 `Unauthorized`）。
    前提是登出請求真的抵達 Supabase：離線時、或送出時 access token 已過期的登出，伺服器端的
    session 沒有被終止，殘留的 token 在過期前仍然有效（notes.md §1）。
    **註銷帳號**則打本服務的 `DELETE /v1/me`；**基本資料**（暱稱、頭像、登入方式…）是 `GET /v1/me`。

    **App 設定**（v4.2.0 起）：`GET /v1/app/update` 更新檢查（強制／有新版由伺服器判定）、
    `POST /v1/app/flags` 功能總開關（逐 key 回布林）。**兩支都不需要認證**——App 一啟動、
    還沒匿名登入或 token 已過期時就能打；`Authorization` 帶了也會被忽略，永遠不會因為 token 而 401。

    **業務請求只帶一個 header**：`Authorization: Bearer <access_token>`（來自匿名登入）。
    apikey 只有 Supabase 的 auth 路徑需要，不再是業務 API 的一部分。

    **錯誤契約**：RFC 9457 problem+json（`Content-Type: application/problem+json`）。
    業務錯誤帶 `code`（凍結 token，enum 見 `Problem` schema）＋有語意的 HTTP 狀態碼
    （400／401／403／404／409／422／429）；**型別／格式錯誤是 400 且沒有 `code`**。
    判斷規則因此只有一條：**有 `code` 才是業務錯誤**——`401` 一律代表 session 問題、
    走刷新流程，不必對 body 做任何比對。

    Postman：直接 import 本檔即可產生 collection；可執行的完整流程見 `postman/`。
servers:
...
    description: >-
...
paths:
...
  /v1/me:
    get:
      tags: [account]
      operationId: me
      summary: 基本資料（頭像、暱稱、登入方式、上次登入時間、會員等級）
      description: |
        呼叫者自己的基本資料，對象只來自 Bearer token 的 `sub`；沒有參數。
        六個鍵**永遠都在**，沒有值的是 `null`（不是缺鍵）。

        - `loginMethod`：`anonymous`（訪客，尚未綁定）、`google`、`apple`。綁了兩個以上取**最早綁定**
          的那個（與 Supabase 自己算 `app_metadata.provider` 的規則相同）。日後新增登入方式＝新增值，
          client 遇到不認得的值走預設分支（notes.md §9）。
        - `nickname`／`avatarUrl`：綁定身分的提供者給的（Google 有姓名與頭像；Apple 沒有頭像、
          姓名也不在 id_token 裡），訪客兩者皆 `null`。本服務沒有編輯端點；經 Supabase 改了
          user metadata 會反映在這裡（使用者自己可控的字串，只回給本人、伺服器不驗證）。
        - `lastLoginAt`：Supabase 記錄的上次登入時間（`last_sign_in_at`）；**訪客一律 `null`**。
        - `memberLevel`：目前恆為 `free`，預留給日後的訂閱等級；新增值為非破壞性變更。
      responses:
        "200":
          description: 基本資料
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      tags: [account]
      operationId: deleteMe
      summary: 註銷帳號（立即、永久、不可逆）
      description: |
        永久刪除呼叫者的帳號：**立即生效、無反悔期、無法復原**。匿名帳號與已綁定帳號同樣適用。
        刪除的對象只來自 Bearer token 本身——沒有 request body、沒有 path 參數，刪不到別人。

        一併消失的：該帳號的全部 session（手上所有 token 立即 401）與綁定身分，以及
        **這個旅人寫下的全部便條**——地圖上未撿的、旅遊紀錄、以及**已被其他旅人撿進收藏的**
        （那些便條從撿藏者的 `/v1/me/collection` 靜默消失，列表只是少一件）。
        **不受影響的**：這個旅人撿過的別人便條——仍留在原作者的 `/v1/me/notes` 裡、仍是已撿走
        （`pickedUpAt` 不變），不會回到地圖上。

        **重試**：回應在網路上遺失（timeout、斷線）時可以直接重試。重試得到 204 ＝已註銷；
        得到 **401** 就照常走刷新流程——註銷成功過的話 session 已不存在、刷新會失敗，
        那才代表已註銷（401 本身也可能只是 token 剛好過期，而第一次請求根本沒送到）。
        註銷不需要自己的「已刪除」錯誤分支。
      responses:
        "204":
          description: 已註銷。沒有 body。
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          description: |
            伺服器端故障，可稍後重試。problem+json、**沒有 `code`**（不是業務錯誤）。
            通常代表帳號**沒有**被刪（同一張 token 仍可正常使用），但有一個窗口例外：
            刪除已經生效、回應卻在回程遺失。重試照 `DELETE /v1/me` 的重試規則判斷即可。
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: about:blank
                status: 500
                title: Internal Server Error
  /v1/app/update:
    get:
      tags: [app]
      operationId: checkUpdate
      summary: 更新檢查（強制更新／有新版，由伺服器判定）
      description: |
        把 App 目前的版本交給伺服器判定：低於**最低相容版本**＝`forceUpdate: true`（必須更新才能
        繼續使用）；低於**最新版**＝`updateAvailable: true`（可提示、不強制；`forceUpdate` 為 true 時
        它必然也是 true）。兩個門檻原樣回傳只供顯示——**client 不要自己比版本**，調門檻不需要發版。

        `version`：1–4 段純數字以點分隔（`1.2`、`1.2.3`、`1.2.3.4`），伺服器逐段按數值比較、
        缺段補 0（`1.2` ＝ `1.2.0`，`1.9` < `1.10`）。缺少或格式不合（`v1.2`、`1.2.0-beta`、空字串）
        → 400 且**沒有 `code`**。目前只有 iOS 一個平台，沒有平台參數；日後加平台為非破壞性變更。

        **不需要認證**：App 啟動第一件事就能打，不必先匿名登入；`Authorization` 帶了也會被忽略
        （過期或已登出的 token 不會讓這支 401）。
      security: []
      parameters:
        - name: version
          in: query
          required: true
          schema:
            type: string
            pattern: '^\d{1,9}(\.\d{1,9}){0,3}$'
            example: "1.2.0"
          description: App 目前的版本（iOS 的 CFBundleShortVersionString）
      responses:
        "200":
          description: 判定結果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateCheck"
        "400":
          $ref: "#/components/responses/FormatBadRequest"
  /v1/app/flags:
    post:
      tags: [app]
      operationId: featureFlags
      summary: 功能總開關（逐 key 回 true／false）
      description: |
        把想知道的開關 key 一次送來，伺服器逐 key 回布林。**伺服器沒設定的 key 一律 `false`**
        （不是錯誤、也不會缺鍵），所以可以放心問還不存在的 key——client 端的預設值也應該是「關」。
        重複的 key 合併。開關是**全站**的，不分使用者；判定在伺服器，撥動不需要發版。

        key 由雙方約定、不寫進本契約（它們是設定不是契約）；命名建議 `區域.功能`（如 `map.heatmap`）。
        字元集 `A-Za-z0-9_.-`、1–64 字元、一次最多 50 個；不合（含 `keys` 缺少、不是字串陣列）
        → 400 且**沒有 `code`**。回應包在 `flags` 物件裡而不是裸 map：日後加欄位不是破壞性變更。

        **不需要認證**（同 `/v1/app/update`）：`Authorization` 帶了也會被忽略。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlagsRequest"
            example: { keys: [map.heatmap, notes.drop] }
      responses:
        "200":
          description: 每個問到的 key 一個布林
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlagsResult"
              example: { flags: { map.heatmap: true, notes.drop: false } }
        "400":
          $ref: "#/components/responses/FormatBadRequest"
...
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        匿名登入取得的 access_token（Supabase GoTrue 簽發）。
        業務請求**只需要這一個 header**；服務只驗它，不代理任何 auth 路徑。
...
  responses:
    Unauthorized:
      description: |
        `not_authenticated`——token 缺失、過期、簽章不符、`sub` 缺失或不是 UUID、
        `aud` 不符、`iss` 不符或缺失、缺 `exp`，以及 **session 已終止**：token 簽得過、
        也還沒過期，但它所屬的 session 已不存在（已登出，或帳號已註銷）。全部同一個答案。
        **401 一律代表 session 問題**：走刷新流程，不需要（也不應該）比對 body。
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: about:blank
            status: 401
            title: not_authenticated
            code: not_authenticated
    FormatBadRequest:
      description: |
        型別／格式錯誤（缺必填、型別不對、格式不合）。problem+json、**沒有 `code`**——不是業務錯誤。
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: about:blank
            status: 400
            title: Bad Request
...
  schemas:
...
    Profile:
      type: object
      description: 呼叫者的基本資料（`GET /v1/me`）。六鍵永遠都在，沒有值的是 null。
      required: [id, nickname, avatarUrl, loginMethod, lastLoginAt, memberLevel]
      properties:
        id:
          type: string
          format: uuid
          description: 使用者 id（＝ token 的 `sub`；綁定帳號前後不變）。只回給本人。
        nickname:
          type: string
          nullable: true
          example: Kevin Chen
          description: 綁定身分提供的姓名；訪客、或提供者沒給（Apple）時為 null
        avatarUrl:
          type: string
          format: uri
          nullable: true
          example: https://lh3.googleusercontent.com/a/default-user
          description: 綁定身分提供的頭像網址（Google 有、Apple 沒有）；訪客為 null
        loginMethod:
          type: string
          example: anonymous
          description: |
            登入方式：`anonymous`（訪客）、`google`、`apple`。綁了多個取最早綁定的那個。
            日後新增值為非破壞性變更（client 走預設分支）。
        lastLoginAt:
          type: string
          format: date-time
          pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}\.\d{6}Z$'
          nullable: true
          example: "2026-10-01T02:03:04.123456Z"
          description: 格式同 Timestamp；Supabase 記錄的上次登入時間（token 刷新不算）。**訪客一律 null**。
        memberLevel:
          type: string
          example: free
          description: 會員等級，目前恆為 `free`（預留給訂閱）；日後新增值為非破壞性變更，不認得的當 `free`。
    UpdateCheck:
      type: object
      description: 更新檢查的判定（`GET /v1/app/update`）。兩個判定給流程用，兩個門檻只供顯示。
      required: [forceUpdate, updateAvailable, minimumVersion, latestVersion]
      properties:
        forceUpdate:
          type: boolean
          description: 低於最低相容版本——必須更新才能繼續使用
        updateAvailable:
          type: boolean
          description: 低於最新版——可提示、不強制。forceUpdate 為 true 時必然也是 true
        minimumVersion:
          type: string
          example: "1.2.0"
          description: 最低相容版本（顯示用；不要拿來自己比）
        latestVersion:
          type: string
          example: "1.4.0"
          description: 最新版（顯示用；不要拿來自己比）
    FlagsRequest:
      type: object
      description: 功能總開關的請求（`POST /v1/app/flags`）
      required: [keys]
      properties:
        keys:
          type: array
          maxItems: 50
          description: 想問的開關 key；0–50 個，重複合併
          items:
            type: string
            pattern: '^[A-Za-z0-9_.-]{1,64}$'
            example: map.heatmap
    FlagsResult:
      type: object
      description: 功能總開關的回應。包在 `flags` 裡（不是裸 map）：日後加欄位才不是破壞性變更
      required: [flags]
      properties:
        flags:
          type: object
          description: 問到的每個 key 一個布林；伺服器沒設定的 key 為 false
          additionalProperties:
            type: boolean
          example: { map.heatmap: true, notes.drop: false }
...
    Problem:
      type: object
      description: |
        RFC 9457 problem+json（`Content-Type: application/problem+json`）。
        **`code` 是唯一的判斷依據**：有 `code` ⇒ 業務錯誤，依 token 決定行為；
        沒有 `code`（型別／格式錯誤、平台層錯誤）⇒ 通用錯誤路徑。
        HTTP 狀態碼提供第一層分流（401 走 session 刷新、429 走重試…），
        業務層再 switch `code`。
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
          default: about:blank
          example: about:blank
          description: >-
            問題型別。本契約目前恆為 `about:blank`（型別由 `code` 表達）。
            日後若改為具體 URI，是非破壞性變更——client 不得依賴它。
        title:
          type: string
          example: too_far
          description: >-
            人類可讀的摘要。業務錯誤時等於 `code`，其餘為通用描述。
            **給人看的，不是判斷依據**——不得對它做字串比對。
        status:
          type: integer
          example: 403
          description: HTTP 狀態碼，與 response 的狀態碼相同（RFC 9457 的慣例）。
        code:
          type: string
          description: |
            凍結的機器碼。**只有業務錯誤才有這個鍵**；型別／格式錯誤的 400
            沒有它（v3 的「第二層閘門」在 v4 就是「有沒有 `code`」）。
            token 字串永久凍結：新增為非破壞性變更（依未知值政策走 default 分支），
            改名或刪除為破壞性變更。
          enum:
            - not_authenticated
            - invalid_coordinates
            - invalid_cursor
            - content_empty
            - content_too_long
            - invalid_style_code
            - invalid_audience
            - active_note_limit
            - private_note_limit
            - note_not_found
            - note_taken
            - own_note
            - too_far
            - pickup_rate_limited
        details:
          type: object
          nullable: true
          additionalProperties: true
          example: { distanceM: 87 }
          description: |
            選配的附帶資料，**v4 起是真的 JSON 物件**（v3 是內容為 JSON 的字串，
            需二次解析——那個規則消失了）。沒有附帶資料時**這個鍵省略**；
            client 不得依賴其存在，也不得區分「沒有這個鍵」與「值為 null」
            （它讓文案更精確，不是流程的前提）。
            目前附帶資料的 token 與內容：
            `too_far` → `{"distanceM": 87}`（伺服器當下算出的實際距離，公尺；與
            NearbyHint 的 `distanceM` 同一算法）；
            `content_too_long` → `{"maxChars": 500}`（Unicode code point）；
            `active_note_limit` → `{"maxActiveNotes": 50}`（未撿**公開**便條的數量上限）；
            `private_note_limit` → `{"maxPrivateNotes": 5000}`（旅遊紀錄的絕對總量上限）；
            `pickup_rate_limited` → `{"retryAfterS": 1800}`（伺服器算出的建議重試秒數，
            與 `Retry-After` header 同值）。
            其餘 token 沒有 `details`。新增鍵、或讓更多 token 附帶 `details`，
            皆為非破壞性變更。
security:
  - Bearer: []